web_backEnd/src/ShopWeb.Tests/Service/ReviewServiceTests.cs
using System;
using System.Collections.Generic;
using System.Threading.Tasks;
using AutoMapper;
using Moq;
using ShopWeb.Core.src.Entity;
using ShopWeb.Core.src.RepoAbstract;
using ShopWeb.Service.src.DTOs;
using ShopWeb.Service.src.Services;
using Xunit;

namespace ShopWeb.Service.Tests
{
    public class ReviewServiceTests
    {
        private readonly ReviewService _reviewService;
        private readonly Mock<IReviewRepository> _reviewRepositoryMock;
        private readonly Mock<IMapper> _mapperMock;

        public ReviewServiceTests()
        {
            _reviewRepositoryMock = new Mock<IReviewRepository>();
            _mapperMock = new Mock<IMapper>();

            _reviewService = new ReviewService(_reviewRepositoryMock.Object, _mapperMock.Object);
        }

        [Fact]
        public async Task CreateReviewAsync_ValidReviewCreateDtoAndUserId_ReturnsTrue()
        {
            // Arrange
            var reviewCreateDto = new ReviewCreateDto
            {
                // Set up review create DTO
            };
            var userId = Guid.NewGuid();
            var review = new Review
            {
                // Set up review
            };

            _mapperMock.Setup(x => x.Map<Review>(reviewCreateDto)).Returns(review);
            _reviewRepositoryMock.Setup(x => x.AddReviewAsync(review)).ReturnsAsync(review);

            // Act
            var result = await _reviewService.CreateReviewAsync(reviewCreateDto, userId);

            // Assert
            Assert.True(result);
            // Add additional assertions as needed
        }

        [Fact]
        public async Task DeleteReviewAsync_ValidReviewIdAndUserId_ReturnsTrue()
        {
            // Arrange
            var reviewId = Guid.NewGuid();
            var userId = Guid.NewGuid();

            _reviewRepositoryMock.Setup(x => x.DeleteReviewAsync(reviewId)).ReturnsAsync(true);

            // Act
            var result = await _reviewService.DeleteReviewAsync(reviewId, userId);

            // Assert
            Assert.True(result);
        }

        [Fact]
        public async Task GetReviewsByProductIdAsync_ValidProductId_ReturnsReviews()
        {
            // Arrange
            var productId = Guid.NewGuid();
            var reviews = new List<Review>
            {
                // Set up list of reviews
            };
            _reviewRepositoryMock.Setup(x => x.GetReviewByProductIdAsync(productId)).ReturnsAsync(reviews);
            _mapperMock.Setup(x => x.Map<IEnumerable<ReviewReadDto>>(reviews)).Returns(new List<ReviewReadDto>());

            // Act
            var result = await _reviewService.GetReviewsByProductIdAsync(productId);

            // Assert
            Assert.NotNull(result);
            // Add additional assertions as needed
        }

        [Fact]
        public async Task GetReviewsByUserIdAsync_ValidUserId_ReturnsReviews()
        {
            // Arrange
            var userId = Guid.NewGuid();
            var reviews = new List<Review>
            {
                // Set up list of reviews
            };
            _reviewRepositoryMock.Setup(x => x.GetReviewByUserIdAsync(userId)).ReturnsAsync(reviews);
            _mapperMock.Setup(x => x.Map<IEnumerable<ReviewReadDto>>(reviews)).Returns(new List<ReviewReadDto>());

            // Act
            var result = await _reviewService.GetReviewsByUserIdAsync(userId);

            // Assert
            Assert.NotNull(result);
            // Add additional assertions as needed
        }

        [Fact]
        public async Task UpdateReviewAsync_ValidReviewUpdateDtoAndId_ReturnsTrue()
        {
            // Arrange
            var reviewUpdateDto = new ReviewUpdateDto
            {
                // Set up review update DTO
            };
            var reviewId = Guid.NewGuid();
            var existingReview = new Review
            {
                // Set up existing review
            };

            _reviewRepositoryMock.Setup(x => x.GetReviewByReviewIdAsync(reviewId)).ReturnsAsync(existingReview);

            // Act
            var result = await _reviewService.UpdateReviewAsync(reviewUpdateDto, reviewId);

            // Assert
            Assert.True(result);
            // Add additional assertions as needed
        }
    }
}
